add multi-agent inference code
