+++ multi-agent/README.md
@@ -1,0 +1,8 @@
+# Multi-agent experiment results
+
+## Result
+The evaluation results are based on a random sample of 50 data samples to reduce API usage costs.
+| Model | Error Flag Accuracy | Error Sentence Detection Accuracy | ROUGE1 | BERTSCORE | BLEURT | AggregateComposite | AggregateScore |
+|-----|---|---|---|---|---|---|---|
+| GPT-5.2 | 0.68 | 0.6 | 0.5860 | 0.6562 | 0.6478 | 0.5024 | 0.6300 |
+| GPT-4.1 | 0.74 | 0.68 | 0.6246 | 0.7082 | 0.6693 | 0.6070 | 0.6674 |
