+++ multi-agent/outputs/medec-ms_gpt-5.2_multi-agent_test_20260206_152335.txt
@@ -1,0 +1,50 @@
+ms-test-17 1 4 "Mycobacterium tuberculosis is suspected."
+ms-test-22 1 2 "The child is suspected to have severe combined immunodeficiency syndrome after examination of his serum shows decreased levels of IgG and a CT scan reveals the absence of a thymus."
+ms-test-29 1 21 "Diffuse cutaneous systemic sclerosis is suspected."
+ms-test-43 1 6 "Hemoglobin 14 g/dL"
+ms-test-45 1 4 "The patient is diagnosed with polyarteritis nodosa."
+ms-test-53 1 13 "The patient is diagnosed with Zollinger-Ellison syndrome."
+ms-test-67 1 15 "The patient is diagnosed with Alzheimer dementia after she cannot recall any of 3 words after 5 minutes."
+ms-test-68 1 17 "The patient is diagnosed with delirium."
+ms-test-74 1 6 "The patient is diagnosed with liposarcoma after histopathologic evaluation of the mass, shown in the image below, confirms malignancy."
+ms-test-82 1 3 "Despite these behaviors, he is meeting his language goals for his age (using 3-word sentences)."
+ms-test-92 1 11 "The patient's behavior is also suspicious for avoidant personality disorder."
+ms-test-95 0 -1 NA
+ms-test-98 1 5 "The patient is diagnosed with delirium after examination."
+ms-test-104 1 15 "HCO3- 40 mEq/L"
+ms-test-107 1 7 "The patient is diagnosed with alpha-1 antitrypsin deficiency after laboratory testing is significant for the following:"
+ms-test-120 1 1 "She states that she has a hard time lifting both her arms, but that they function normally otherwise."
+ms-test-121 1 6 "The patient is diagnosed with trichomoniasis after her pelvic exam; she has vulvovaginal erythema and a \"strawberry cervix\" that is tender to motion, with minimal green-yellow malodorous discharge."
+ms-test-130 0 -1 NA
+ms-test-134 0 -1 NA
+ms-test-142 0 -1 NA
+ms-test-178 0 -1 NA
+ms-test-183 1 7 "The patient is diagnosed with aortic regurgitation."
+ms-test-207 0 -1 NA
+ms-test-217 1 7 "The patient is empirically started on 3 units of uncrossmatched O-negative blood and pantoprazole."
+ms-test-234 1 2 "The patient is diagnosed with X-linked agammaglobulinemia after lab investigations reveal an undetectable level of all serum immunoglobulins."
+ms-test-242 1 7 "The patient is diagnosed with advanced sleep phase disorder."
+ms-test-254 1 2 "After a pericardial window is performed, while observing sinus tachycardia on the patient's telemetry, you note that the radial pulse cannot be palpated during inspiration."
+ms-test-265 1 10 "Folate is administered; however, on follow-up, her laboratory parameters show no interval change in her MCV or platelet level."
+ms-test-267 0 -1 NA
+ms-test-288 0 -1 NA
+ms-test-305 1 4 "The patient is started on combined vasoactive and endoscopic therapy after he is noted to have jaundice, splenomegaly, and multiple spider angiomas over his chest."
+ms-test-331 1 4 "The lesion is diagnosed as basal cell carcinoma, and the patient undergoes standard excision."
+ms-test-344 1 8 "The patient is scheduled for a modified radical mastectomy followed by adjuvant radiation therapy."
+ms-test-352 1 5 "Hydrocortisone is administered to ensure normal functioning of her hypothalamic-pituitary-adrenal (HPA) axis."
+ms-test-358 0 -1 NA
+ms-test-380 0 -1 NA
+ms-test-386 0 -1 NA
+ms-test-391 0 -1 NA
+ms-test-414 1 10 "High-dose intravenous methylprednisolone is prescribed."
+ms-test-444 1 7 "He is prescribed an appropriate medication called cimetidine for his symptoms and is told to follow up in 2 weeks."
+ms-test-454 1 7 "Start an SSRI."
+ms-test-464 1 5 "On physical examination, she has a grade 2/6 holosystolic murmur heard best at the left lower sternal border, bilateral bibasilar crackles on lung auscultation, and a normal abdominal examination."
+ms-test-465 1 7 "Hemoglobin: 11 g/dL"
+ms-test-492 1 4 "She gives a printed list of such drugs, including efavirenz and delavirdine, to the woman for educational and informational purposes."
+ms-test-494 1 9 "The physician says he may prescribe varenicline for the patient after a full cardiac workup is performed."
+ms-test-503 1 2 "The patient is given azithromycin after a cold agglutination test conducted on her blood shows clumping when placed at 4 degrees Celsius."
+ms-test-504 0 -1 NA
+ms-test-525 1 21 "Potentially toxic drugs are discontinued."
+ms-test-570 1 6 "Oral metronidazole is prescribed after the gynecologic examination shows increased production of a white-yellow vaginal discharge."
+ms-test-585 0 -1 NA
